requirements.txt
# Core Framework
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# AI & ML Libraries
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
shap>=0.44.0
imbalanced-learn>=0.12.0

# Google AI Integration
google-generativeai>=0.3.0
google-api-python-client>=2.100.0

# Data Visualization
plotly>=5.15.0
seaborn>=0.12.0
matplotlib>=3.7.0
wordcloud>=1.9.0
networkx>=3.1.0

# Text Processing & NLP
nltk>=3.8.0
spacy>=3.7.0
textblob>=0.17.0
sastrawi>=1.0.0
transformers>=4.30.0
torch>=2.0.0
tensorflow>=2.13.0
gensim>=4.3.0
langdetect>=1.0.0
polyglot>=16.7.0

# Text Analysis
vaderSentiment>=3.3.0
yake>=0.4.0
keybert>=0.8.0
bertopic>=0.16.0

# Data Export & Reports
fpdf2>=2.7.0
reportlab>=4.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
joblib>=1.3.0
nbformat>=5.7.0
jupyter>=1.0.0

# Database & Storage
sqlalchemy>=2.0.0
# sqlite3 is built into Python, no need to install
pymongo>=4.5.0
redis>=5.0.0

# Web & API
requests>=2.30.0
httpx>=0.24.0
fastapi>=0.100.0
uvicorn>=0.20.0
websockets>=11.0
websocket-client>=1.5.0
psutil>=5.9.0

# Utilities & Helpers
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0
tqdm>=4.65.0
colorama>=0.4.0
rich>=13.5.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Statistical Analysis
scipy>=1.10.0
statsmodels>=0.14.0
pingouin>=0.5.0

# Time Series Analysis
prophet>=1.1.0
tslearn>=0.6.0

# Geospatial Analysis
geopandas>=0.13.0
folium>=0.14.0

# Performance & Optimization
numba>=0.57.0
cython>=3.0.0
dask>=2023.9.0

# Testing & Quality
pytest>=7.3.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Security & Authentication
cryptography>=41.0.0
passlib>=1.7.0
python-jose>=3.3.0
bcrypt>=4.0.0

# Monitoring & Logging
loguru>=0.7.0
sentry-sdk>=1.30.0

# Deployment
gunicorn>=21.0.0
psycopg2-binary>=2.9.0

# Additional Tools
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1.0
chardet>=5.2.0